cmd/link: Fix archreloc to generate correct relocations for pcrel instructions
When auipcImm has its MSB set, it gets sign extended to 64 bits, which means the
most significant 32 bits of val always get set essentially ignoring the ADDI instruction
present in second.
This patch fixes this bug by interpreting auipcImm (and secondImm) as unsigned integers.

+++ src/cmd/link/internal/riscv/asm.go
@@ -130,8 +130,8 @@ func archreloc(ctxt *ld.Link, r *ld.Reloc, s *ld.Symbol, val *int64) int {
 		auipc := int64(uint32(*val))
 		second := int64(uint32(*val >> 32))
 
-		auipc = (auipc &^ riscv.UTypeImmMask) | auipcImm
-		second = (second &^ secondImmMask) | secondImm
+		auipc = (auipc &^ riscv.UTypeImmMask) | int64(uint32(auipcImm))
+		second = (second &^ secondImmMask) | int64(uint32(secondImm))
 
 		*val = second<<32 | auipc
